+++ public_html/ckeditor/plugins/fileman/js/custom.js
@@ -1,5 +1,5 @@
 /*
-  RoxyFileman - web based file manager. Ready to use with CKEditor, TinyMCE. 
+  RoxyFileman - web based file manager. Ready to use with CKEditor, TinyMCE.
   Can be easily integrated with any other WYSIWYG editor or CMS.
 
   Copyright (C) 2013, RoxyFileman.com - Lyubomir Arsov. All rights reserved.
@@ -19,10 +19,11 @@
 
   Contact: Lyubomir Arsov, liubo (at) web-lobby.com
 */
+
 function FileSelected(file){
   /**
    * file is an object containing following properties:
-   * 
+   *
    * fullPath - path to the file - absolute from your site root
    * path - directory in which the file is located - absolute from your site root
    * size - size of the file in bytes
@@ -31,7 +32,30 @@ function FileSelected(file){
    * ext - file extension
    * width - if the file is image, this will be the width of the original image, 0 otherwise
    * height - if the file is image, this will be the height of the original image, 0 otherwise
-   * 
+   *
+   */
+  // Get the ID of the input to fill
+  var fieldId = RoxyUtils.GetUrlParam('txtFieldId');
+  $(window.parent.document).find('#' + fieldId+'_img').attr('src', file.fullPath);
+  $(window.parent.document).find('#' + fieldId).attr('value', file.fullPath);
+//  $(window.parent.document).find('#' + fieldId).attr('value', file.fullPath);
+//  $(window.parent.document).find('#' + fieldId).value = file.fullPath;
+//  $('input.'+fieldId).val(File.fullPath);
+  window.parent.closeStoryImage();
+}
+function FileSelectedxx(file){
+  /**
+   * file is an object containing following properties:
+   *
+   * fullPath - path to the file - absolute from your site root
+   * path - directory in which the file is located - absolute from your site root
+   * size - size of the file in bytes
+   * time - timestamo of last modification
+   * name - file name
+   * ext - file extension
+   * width - if the file is image, this will be the width of the original image, 0 otherwise
+   * height - if the file is image, this will be the height of the original image, 0 otherwise
+   *
    */
   alert('"' + file.fullPath + "\" selected.\n To integrate with CKEditor or TinyMCE change INTEGRATION setting in conf.json. For more details see the Installation instructions at http://www.roxyfileman.com/install.");
 }
@@ -40,6 +64,6 @@ function GetSelectedValue(){
   * This function is called to retrieve selected value when custom integration is used.
   * Url parameter selected will override this value.
   */
-  
+
   return "";
 }